requirements.txt
google-cloud-bigquery
requests
python-dotenv
loguru